# [GPU][MTL] Resolve long token performance regression in MTL 125H plat…