@yeonbok 2024-12-20
Hi @riverlijunjie according to @isanghao , PR26940 should not affect MTL because the PR was for onednn case. Could you clarify how PR26940 affected perf?

(And apart from the question, I think we can apply this change though)