@riverlijunjie 2024-12-23
> Hi @riverlijunjie according to @isanghao , PR26940 should not affect MTL because the PR was for onednn case. Could you clarify how PR26940 affected perf?
> 
> (And apart from the question, I think we can apply this change though)

PR26940 updates dynamic_quantize for cldnn, it should take effect on MTL, am i right?