@peterchen-intel 2025-01-03
Replaced by https://github.com/openvinotoolkit/openvino/pull/28209